+++ examples/create_rundir.sh
@@ -1,0 +1,3 @@
+#!/bin/bash
+
+python create_rundir.py c12_config.yml example_rundir